X86SMEP pass
This commit adds an X86_64 Supervisor Mode Execution Prevention Pass.
The X86SMEP pass instruments indirect branches and return instructions
with code that verifies control will not flow to user space memory.

TODO: Currently, the pass adds a nop instruction instead of a trap
for invalid flows. This is because the linux kernel has many parts
that execute in user mode which should not be instrumented with this pass.
Until we identify these components, we emulate the cost of the intrumentation
with a NOP instruction instead.