Skeleton X86KernelShadowStack pass
This commit adds an empty x86_64 pass that will be used to implement
a shadow call stack for the Linux kernel (version 5.7) using PKK
(i.e., Intel's PKU feature enabled for kernel).

The basic idea for the kernel shadow stack is the following:
Linux Kernel (v5.7) uses 16KB (4 pages) for thread stacks. With minor
modifications, it is possible to double each stack to 32KB and use half of
it as shadow stack for return addresses. The shadow stack will be write-protected
by default using PKK.
The address of the top of the shadow stack can be easily found by
just flipping bit 15 of the current %rsp value.