Race-free Shadow Stack Patch
The commit instruments all callsites as well as the prolog
and epilog of every function in a module. In IskiOS's shadow
stack, we maintain the following invariant: on entry to a function,
%r10 contains the return address. On function return, %r10 contains
the return address and we jump through %r10.
Since the return address is passed through a register,
there's no opportunity for race between CALL and function entry.
In addition, since we jump through %r10, there's no opportunity
for race between ret instruction and actual return.

Function Prolog Instrumentation:

The added instructions compute the address of the shadow stack,
enable write access to shadow stack, store the address
saved in %r10 (i.e., the return address) to the top of the
shadow stack, and disable the write access.

Function Epilog Instrumentation:

The added instructions copy the return address from the top
of the shadow stack to %r10 and jump to that return address.

Callsite Instrumentation:

Before the CALL instruction, the pass adds a LEA instruction
which loads the effective address of the return site to %r10.

Note that this shadow stack needs CONFIG_ISKIOS_SHADOW_STACK to
be enabled in the IskiOS kernel.