Safe WRPKRU Patch
This commit adds a SAFE_WRPKRU option in the X86KernelShadowStack pass.
When enabled, all wrpkru instructions are instrumented with code that
checks and verifies that execution is taking place in kernel mode.
This happens by checking the last byte of the code segment register.
If the current privilege level (CPL) is not 3 when wrpkru is executed,
then the kernel will trap. This indicates that user mode code is
attempting to bypass IskiOS's protection by changing access rights
via the wrpkru instruction.