Shadow Write Optimization (SWO)
This commit adds SWO option to the X86ShadowStack pass.
When enabled, the function prologs are modified in the following way:

Before the rdpkru/wrpkru instructions are executed, an additional check
verfies that the return address on top of the shadow stack is different
from the current value of %rsp. If the two values are the same, then
there's no need for re-writing the return address to the shadow stack.

This option improves performance as reading from our PKU-protected
shadow-stack is cheap, while writing to it requires 2 wrpkru instructions
which are expensive. This will be especially beneficial for tail-calls.
Note that the two additional instructions may slow non tail-call cases.